+++ source/extensions/omni.isaac.ui/omni/isaac/ui/element_wrappers/__init__.py
@@ -1,2 +1,13 @@
-from .base_ui_element_wrappers import UIWidgetWrapper, UIFrameWrapper
-from .ui_widget_wrappers import Button, StateButton, DropDown, FloatField
+from .base_ui_element_wrappers import UIWidgetWrapper
+from .ui_widget_wrappers import (
+    FloatField,
+    IntField,
+    StringField,
+    CheckBox,
+    DropDown,
+    CollapsableFrame,
+    TextBlock,
+    ColorPicker,
+    Button,
+    StateButton,
+)